Update gen-hcce to have parity with render_hcce.sh
Update the input-values.yaml config file to match what's in render_hcce.sh
Generate a new key pair for the certificate.
Only modify the common name for the certificate.
Preprocess the input-values.yaml config file to replace the database username and password variables in PGRST_DB_URI and PSQL with their actual values.
Skip saving out any of the keys or certificates and just use them directly where needed.
Skip saving things in environment variables and just use them directly where needed.
Add base64 encoded versions of the certificate and its private key to the config so that they can be transferred properly to the end hcce.yaml file.
Fix the replacePlaceholders function so that it replaces all variables that have a corresponding entry in the config file, even if that entry's value is false.