Fix the subdomain generation for the gen-ssl script.
Output the correct URL to the cbb.yaml file for each SSL certificate being generated instead of just the URL for the root domain.  Outputting only the root domain caused the script to only generate the certificate for the root domain, and so none of the certificates for the subdomains were generated causing the script to fail when it attempted to get the subdomain certificates after it thought it had finished generating them.